Move download files to separate folder
